+++ CHANGELOG.md
@@ -31,5 +31,6 @@
 
 - Support markdown descriptions for collectibles and bios
 - Fixup ID setup not updating addresses
+- Fixup collection contract owner bug
 - Fix 404 errors incorrectly displayed on Orbit
 - Custom 500 error page